+++ packages/plugins/Trader/src/SiteAdaptor/trader/TraderInfo.tsx
@@ -1,16 +1,16 @@
 import { memo, useMemo } from 'react'
 import { useAsyncRetry } from 'react-use'
-import { type ChainId, formatWeiToEther } from '@masknet/web3-shared-evm'
-import { resolveTradeProviderName } from '../../helpers/resolveTradeProviderName.js'
-import { NetworkPluginID } from '@masknet/shared-base'
-import { multipliedBy, formatBalance, ZERO, formatCurrency, formatPercentage } from '@masknet/web3-shared-base'
-import { PluginTraderRPC } from '../../messages.js'
-import { TradeProvider } from '@masknet/public-api'
-import { useGreatThanSlippageSetting } from './hooks/useGreatThanSlippageSetting.js'
-import { useChainContext, useNativeTokenPrice, useNetworkContext, useWeb3Others } from '@masknet/web3-hooks-base'
-import { DefaultTraderPlaceholderUI, TraderInfoUI } from './components/TraderInfoUI.js'
-import type { TraderAPI } from '@masknet/web3-providers/types'
 import type { AsyncStateRetry } from 'react-use/lib/useAsyncRetry.js'
+import { type ChainId, formatWeiToEther } from '@masknet/web3-shared-evm'
+import { TradeProvider } from '@masknet/public-api'
+import { NetworkPluginID } from '@masknet/shared-base'
+import type { TraderAPI } from '@masknet/web3-providers/types'
+import { multipliedBy, formatBalance, ZERO, formatCurrency, formatPercentage } from '@masknet/web3-shared-base'
+import { useChainContext, useNativeTokenPrice, useNetworkContext, useWeb3Others } from '@masknet/web3-hooks-base'
+import { PluginTraderRPC } from '../../messages.js'
+import { useGreatThanSlippageSetting } from './hooks/useGreatThanSlippageSetting.js'
+import { resolveTradeProviderName } from '../../helpers/resolveTradeProviderName.js'
+import { DefaultTraderPlaceholderUI, TraderInfoUI } from './components/TraderInfoUI.js'
 
 export interface TraderInfoProps {
     trade: AsyncStateRetry<TraderAPI.TradeInfo>
