+++ packages/plugins/Trader/src/SiteAdaptor/trending/CoinIcon.tsx
