--- packages/plugins/Trader/src/SiteAdaptor/trending/components/index.ts
@@ -1,1 +1,0 @@
-export * from './CoinIcon.js'
