+++ packages/plugins/Trader/src/register.ts
@@ -1,6 +1,6 @@
 import { registerPlugin } from '@masknet/plugin-infra'
-import { base } from './base.js'
 import type { ChainId } from '@masknet/web3-shared-evm'
+import { base } from './base.js'
 
 registerPlugin<ChainId>({
     ...base,
@@ -13,9 +13,5 @@ registerPlugin<ChainId>({
         load: () => import('./Dashboard/index.js'),
         hotModuleReload: (hot) =>
             import.meta.webpackHot?.accept('./Dashboard', () => hot(import('./Dashboard/index.js'))),
-    },
-    Worker: {
-        load: () => import('./Worker/index.js'),
-        hotModuleReload: (hot) => import.meta.webpackHot?.accept('./Worker', () => hot(import('./Worker/index.js'))),
     },
 })
