+++ packages/plugins/Trader/src/trader/uniswap/useTransactionDeadline.ts
@@ -1,9 +1,9 @@
 import { useMemo } from 'react'
 import { BigNumber } from 'bignumber.js'
 import { ChainId } from '@masknet/web3-shared-evm'
-import { DEFAULT_TRANSACTION_DEADLINE, L2_TRANSACTION_DEADLINE } from '../../constants/index.js'
-import { useBlockTimestamp, useChainContext, useNetworkContext } from '@masknet/web3-hooks-base'
 import { NetworkPluginID } from '@masknet/shared-base'
+import { useBlockTimestamp, useChainContext, useNetworkContext } from '@masknet/web3-hooks-base'
+import { DEFAULT_TRANSACTION_DEADLINE, L2_TRANSACTION_DEADLINE } from '../../constants/index.js'
 
 export function useTransactionDeadline() {
     const { chainId } = useChainContext()
