+++ packages/plugins/Trader/src/trader/uniswap/useV2BestTrade.ts
@@ -1,10 +1,10 @@
 import { useMemo } from 'react'
 import { Trade } from '@uniswap/v2-sdk'
 import type { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
+import type { TradeProvider } from '@masknet/public-api'
 import { isTradeBetter } from '../../helpers/isTradeBetter.js'
 import { useAllCommonPairs } from './useAllCommonPairs.js'
 import { BETTER_TRADE_LESS_HOPS_THRESHOLD, MAX_HOP } from '../../constants/index.js'
-import type { TradeProvider } from '@masknet/public-api'
 
 export function useV2BestTradeExactIn(
     tradeProvider: TradeProvider,
